Made edit book input fields writeable
Load value initally from prop to state
